Added client/server side validation to console. Added foreach loop in…

+++ NintendoInventory.DB/dbo/Tables/Console.sql
@@ -1,11 +1,17 @@
 ﻿CREATE TABLE [dbo].[Console] (
-    [ConsoleID]       INT           IDENTITY (1, 1) NOT NULL,
-    [ConsoleName]     VARCHAR (250) NOT NULL,
-    [ReleaseDate]     DATETIME      NULL,
-    [ConsoleImageURL] VARCHAR (300) NOT NULL,
-    [Price]           MONEY         NOT NULL,
+    [ConsoleID]       INT             IDENTITY (1, 1) NOT NULL,
+    [ConsoleName]     VARCHAR (250)   NOT NULL,
+    [ReleaseDate]     DATETIME        NULL,
+    [ConsoleImageURL] VARCHAR (300)   NOT NULL,
+    [Price]           DECIMAL (19, 2) NOT NULL,
     CONSTRAINT [PK_Console] PRIMARY KEY CLUSTERED ([ConsoleID] ASC)
 );
+
+
+
+
+
+
